Replaced duration/timestamp with their new versions
